% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfs.R, R/ipfs_data.R, R/ipfs_info.R,
%   R/ipfs_ls.R, R/ipfs_server.R, R/ipfs_stats.R
\name{ipfs_config}
\alias{ipfs_config}
\alias{ipfs_browse}
\alias{ipfs_webui}
\alias{ipfs_commands}
\alias{ipfs_data}
\alias{ipfs_info}
\alias{ipfs_ls}
\alias{ipfs_server}
\alias{ipfs_stats}
\title{IPFS
Connect to the inter-planetary file system.}
\usage{
ipfs_config()

ipfs_browse(key, gateway = "https://ipfs.io/ipfs/")

ipfs_webui()

ipfs_commands()

ipfs_data(key)

ipfs_info()

ipfs_ls(key)

ipfs_server(new_url = NULL)

ipfs_stats()
}
\arguments{
\item{key}{character with path}

\item{gateway}{any IPFS gateway server}

\item{new_url}{set a different API server. Default is "http://127.0.0.1:5001"}
}
\description{
IPFS
Connect to the inter-planetary file system.

ipfs_data

ipfs_ls

ipfs_server

ipfs_stats
}
