% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfs_download.R
\name{ipfs_download}
\alias{ipfs_download}
\title{ipfs_download}
\usage{
ipfs_download(key, path = NULL)
}
\arguments{
\item{path}{filename to save download.}
}
\description{
ipfs_download
}
\examples{
ipfs_download('QmR7GSQM93Cx5eAg6a6yRzNde1FQv7uL6X1o4k7zrJa3LX/ipfs.draft3.pdf')
# utils::browseURL('ipfs.draft3.pdf')
}
