% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfs_get.R
\name{ipfs}
\alias{ipfs}
\alias{ipfs_get}
\title{ipfs_get}
\usage{
ipfs_get(key)
}
\arguments{
\item{key}{a multihash key}
}
\description{
ipfs_get
}
\examples{
# From 'getting started'
ipfs_daemon(TRUE)
ipfs_info()
ipfs_get('QmYwAPJzv5CZsnA625s3Xf2nemtYgPpHdWEz79ojWnPbdG')
}
\references{
IPFS api: \url{https://ipfs.io/docs/api/}
Draft paper: \url{https://ipfs.io/ipfs/QmR7GSQM93Cx5eAg6a6yRzNde1FQv7uL6X1o4k7zrJa3LX/ipfs.draft3.pdf}
}
