% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfs_json.R
\name{ipfs_json}
\alias{ipfs_json}
\title{ipfs_json}
\usage{
ipfs_json(command, ...)
}
\arguments{
\item{...}{}
}
\value{
character from jsonlite::fromJSON
}
\description{
ipfs_json
}
\examples{
 ipfs_json("object/get", arg = "QmYwAPJzv5CZsnA625s3Xf2nemtYgPpHdWEz79ojWnPbdG")
}
